Coin-M: rename package from coinM to delivery
