+++ delivery/client_test.go
@@ -1,4 +1,4 @@
-package coinM
+package delivery
 
 import (
 	"bytes"
