+++ v2/delivery/exchange_info_service.go
@@ -1,4 +1,4 @@
-package coinM
+package delivery
 
 import (
 	"context"
