+++ v2/delivery/exchange_info_service_test.go
@@ -1,4 +1,4 @@
-package coinM
+package delivery
 
 import (
 	"testing"
