+++ v2/delivery/order_service.go
@@ -1,4 +1,4 @@
-package coinM
+package delivery
 
 import (
 	"context"
